src: dynamically count decibel compensation value for songs
This commit fix #8.
As already explained in the issue, even if ffmpeg wasn't correctly reporting it, compensating songs max_volume value by a fixed decibel amount (as 5.0dB) was bringing lot of mp3 files to overcome a balanced value of 0dB, resulting to a crackle audio.
Dynamically finding the correct value seems to fix the issue.