src: introduce `-fix <filename>` flag
If activated, the application will fetch all the song informations from <filename> and will re-trigger a download for it.
This is needed if downloaded songs were not the ones we were looking for, and we want quickly find a more accurate one (eventually, with `-interactive` flag, too).
This is the only flag that will make SpotiTube work without Spotify at all.