src: introduce -flush-different flag
Similarly to -flush-missing, this flag enables flush of metadata also when there already is a tag on the local file, but it's not up-to-date with the one spotitube has calculated.
Example of use: if you already fetched songs and their lyrics before Genius lyrics provider got integrated and you want to flush all the lyrics trying to get them from Genius, toggle this flag.