src: spotify: add -authenticate-outside flag
By default set to false, it enables authentication socket to be reached
on any interface and set the spotify authentication URL to call
spotitube.local as callback URI.
The user is only asked to set spotitube.local (e.g. using hosts file) name resolution
to the machine which spotitube is being executed on, on the machine the
authetication flow is being followed on.
Needed to fix #42.